// repos/rujira-ui/packages/main/src/strategies/components/__generated__/StakingPoolPageDepositFragment.graphql.ts
export type AssetType = "LAYER_1" | "NATIVE" | "SECURED" | "SYNTH" | "%future added value";
export type Chain = "AVAX" | "BASE" | "BCH" | "BSC" | "BTC" | "DOGE" | "ETH" | "GAIA" | "KUJI" | "LTC" | "NOBLE" | "OSMO" | "SOL" | "TERRA" | "TERRA2" | "THOR" | "TON" | "TRON" | "TRON" | "XRP" | "%future added value";
import { FragmentRefs } from "relay-runtime";
export type StakingPoolPageDepositFragment$data = {
  readonly address: string;
  readonly bondAsset: {
    readonly asset: string;
    readonly chain: Chain;
    readonly metadata: {
      readonly decimals: number;
      readonly symbol: string;
    };
    readonly price: {
      readonly current: bigint | null | undefined;
    } | null | undefined;
    readonly type: AssetType;
    readonly variants: {
      readonly native: {
        readonly denom: string;
      } | null | undefined;
    };
  };
  readonly " $fragmentType": "StakingPoolPageDepositFragment";
};
export type StakingPoolPageDepositFragment$key = {
  readonly " $data"?: StakingPoolPageDepositFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"StakingPoolPageDepositFragment">;
};