// repos/rujira-ui/packages/main/src/strategies/components/__generated__/StakingPoolPageFragment.graphql.ts
export type AprStatus = "AVAILABLE" | "NOT_APPLICABLE" | "SOON" | "%future added value";
export type AssetType = "LAYER_1" | "NATIVE" | "SECURED" | "SYNTH" | "%future added value";
export type Chain = "AVAX" | "BASE" | "BCH" | "BSC" | "BTC" | "DOGE" | "ETH" | "GAIA" | "KUJI" | "LTC" | "NOBLE" | "OSMO" | "SOL" | "TERRA" | "TERRA2" | "THOR" | "TON" | "TRON" | "TRON" | "XRP" | "%future added value";
import { FragmentRefs } from "relay-runtime";
export type StakingPoolPageFragment$data = {
  readonly address: string;
  readonly bondAsset: {
    readonly asset: string;
    readonly chain: Chain;
    readonly metadata: {
      readonly decimals: number;
      readonly symbol: string;
    };
    readonly type: AssetType;
    readonly variants: {
      readonly native: {
        readonly denom: string;
      } | null | undefined;
    };
  };
  readonly id: string;
  readonly revenueAsset: {
    readonly asset: string;
    readonly metadata: {
      readonly symbol: string;
    };
  };
  readonly stakingSummary: {
    readonly apr: {
      readonly status: AprStatus;
      readonly value: bigint | null | undefined;
    };
  };
  readonly status: {
    readonly accountBond: bigint;
    readonly accountRevenue: bigint;
    readonly id: string;
    readonly liquidBondShares: bigint;
    readonly liquidBondSize: bigint;
    readonly pendingRevenue: bigint;
    readonly valueUsd: bigint;
  } | null | undefined;
  readonly " $fragmentSpreads": FragmentRefs<"StakingPoolPageDepositFragment">;
  readonly " $fragmentType": "StakingPoolPageFragment";
};
export type StakingPoolPageFragment$key = {
  readonly " $data"?: StakingPoolPageFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"StakingPoolPageFragment">;
};